+++ src/graphTheory/edgeArray.js
@@ -4,7 +4,6 @@ var Edge = require('./edge');
 // var DirectedEdge = require('./directedEdge');
 var RobustArray = require('./robustArray');
 
-
 /**
  * represents a set of Edges
  * @exports EdgeArray
@@ -12,18 +11,8 @@ var RobustArray = require('./robustArray');
  * @memberOf! module:graphTheory
  */
 class EdgeArray extends RobustArray.SetifyType(Edge) {
-    getNodes() {
-        return this.reduce((prev, next) => prev.union(next.nodes), this[0].nodes);
-        //     return this.reduce((nArray, e) => {
-        //         nArray.unionize(e.nodes);
-        //         return nArray;
-        //     }, new NodeArray());`
-    }
-    getNeighbors(nodeArg) {
-        return this.edgesWithNode(nodeArg).reduce((nArray, e) => nArray.push(e.getNeighbor(nodeArg)), new this[0].nodes.constructor());
-    }
-    edgesWithNode(nodeArg) {
-        return this.filter(e => e.containsNode(nodeArg));
+    edgesWithNode(nArg) {
+        return this.filter(e => e.containsNode(nArg));
     }
     edgeByNodes(n1, n2) {
         return this.find(e => (e.containsNode(n1) && e.containsNode(n2)));
@@ -31,10 +20,13 @@ class EdgeArray extends RobustArray.SetifyType(Edge) {
     edgesByArray(nArr) {
         return nArr.reduce((eArr, nNode) => eArr.unionize(this.edgesWithNode(nNode)), new this.constructor());
     }
-    inferNodeArray(nodes = this[0].nodes) {
-        return new nodes.constructor();
+    getNodes() {
+        return this.reduce((nArr, e) => nArr.union(e.nodes), this[0].nodes);
     }
-    // static injectDependency(NAClass = )
+    getNeighbors(nArg) {
+        return this.edgesWithNode(nArg).reduce((nArray, e) => nArray.unionize(e.nabeArray(nArg)), this[0].nabeArray(nArg));
+    }
+
 }